% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{clean_dupes}
\alias{clean_dupes}
\title{Clean duplicates}
\usage{
clean_dupes(longDF, pid_col = "ID", keep_text = FALSE)
}
\arguments{
\item{longDF}{longDF}

\item{pid_col}{pid_col}

\item{keep_text}{keep_text?}
}
\description{
Clean duplicates
}
