% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{create_key_from_rubric}
\alias{create_key_from_rubric}
\title{Create \code{psych} key from rubric}
\usage{
create_key_from_rubric(rubricsDF, scale_name = NULL)
}
\arguments{
\item{rubricsDF}{rubricsDF}

\item{scale_name}{scale_name}
}
\value{
a key list for use in the \code{psych} package.
}
\description{
Create \code{psych} key from rubric
}
