% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{creds_from_file}
\alias{creds_from_file}
\title{creds_from_file}
\usage{
creds_from_file(creds_yaml = "credentials.yaml")
}
\arguments{
\item{creds_yaml}{The .yaml file that contains the user's credentials}
}
\description{
Reads the API key and base URL from a file and sets up the environment
variables necessary for accessing data. Uses
\code{\link[qualtRics]{qualtrics_api_credentials}}, but does not save
credentials to the .Renviron file.
}
