% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{fix_ids}
\alias{fix_ids}
\title{Fix Participant IDs}
\usage{
fix_ids(x, pid_cols, prefix = NULL, format_ids = FALSE)
}
\arguments{
\item{x}{A data.frame or data.table. Can be the whole table returned by
\code{\link{get_survey_data}}.}

\item{pid_cols}{A character vector of column names in order of priority. x
does not need to include all columns in pid_cols.}

\item{prefix}{A regular expression that will be removed prior to conversion
to numeric.}

\item{format_ids}{is either FALSE or an expression understood by
\code{sprintf} used to format the resulting IDs}
}
\value{
A numeric or character vector of ids.
}
\description{
This function first removes the prefix specified in \code{prefix} (if found)
and then converts all columns in pid_cols to numeric. It then takes the first
non-NA value found, using column order in pid_cols for priority. It returns a
vector of ids, with include_prefix prepended if supplied. This assumes x is a
data.table.
}
