% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api_requests.R
\name{format_request}
\alias{format_request}
\title{Format a list of arguments into a Qualtrics API request.}
\usage{
format_request(api_args,
  rooturl = "https://survey.qualtrics.com/WRAPI/ControlPanel/api.php")
}
\arguments{
\item{api_args}{List of arguments}
}
\description{
Format a list of arguments into a Qualtrics API request.
}

