% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{get_items_in_rubric}
\alias{get_items_in_rubric}
\title{Get items from data file that exist in a rubric}
\usage{
get_items_in_rubric(dataDF, rubricDF)
}
\arguments{
\item{dataDF}{A long-form data.frame as would be returned by \code{\link{get_survey_data}}}

\item{rubricDF}{A rubric data.frame as would be returned by \code{\link{get_rubrics}}}
}
\value{
returns a long-format data.frame with just items appearing in \code{rubricDF}
}
\description{
Get items from data file that exist in a rubric
}
