% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{get_rubrics}
\alias{get_rubrics}
\title{Get rubrics}
\usage{
get_rubrics(rubric_filenames, source = "csv")
}
\arguments{
\item{rubric_filenames}{Vector of file paths named "file".}

\item{source}{Unused, default's to 'csv' for now.}
}
\value{
Returns a long data frame of rubrics with names:
[1] "file"           "data_file_name" "scale_name"     "column_name"
[5] "reverse"        "transform"      "scored_scale"   "include"
[9] "min"            "max"
}
\description{
Get rubrics
}
