% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_requests.R
\name{get_survey_id}
\alias{get_survey_id}
\title{Get the Qualtrics survey id from the survey name}
\usage{
get_survey_id(creds, survey_name, user = creds$user, token = creds$token,
  format = "JSON", version = "2.4")
}
\arguments{
\item{creds}{Qualtrics credentials. A list with values for user and token.}

\item{survey_name}{Qualtris survey name.}

\item{user}{Qualtrics username.}

\item{token}{Qualtrics API token.}

\item{format}{The format of the requested data. Default is JSON.}

\item{version}{The API version to use. Defaults to the latest version, 2.4.}
}
\description{
Get the Qualtrics survey id from the survey name
}
