% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_requests.R
\name{get_survey_responses}
\alias{get_survey_responses}
\title{Return a data.frame of the survey responses for a given survey}
\usage{
get_survey_responses(
  surveyID,
  label = FALSE,
  convert = FALSE,
  force_request = TRUE,
  ...
)
}
\arguments{
\item{surveyID}{Unique survey ID as returned from \code{\link{get_surveys}}.}

\item{label}{default is \code{FALSE}.}

\item{convert}{default is \code{FALSE}.}

\item{force_request}{default is \code{TRUE}.}

\item{...}{See \code{\link[qualtRics]{fetch_survey}.}}
}
\description{
See \code{\link[qualtRics]{fetch_survey}} for complete information on available options.
}
