% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/survey_requests.R
\name{get_surveys}
\alias{get_surveys}
\title{Return a list of all surveys available to the user}
\usage{
get_surveys(creds, user = creds$user, token = creds$token,
  format = "JSON", version = "2.4")
}
\arguments{
\item{creds}{Qualtrics credentials. A list with values for user and token.}

\item{user}{Qualtrics username.}

\item{token}{Qualtrics API token.}

\item{format}{The format of the requested data. Default is JSON.}

\item{version}{The API version to use. Defaults to the latest version, 2.4.}
}
\description{
Return a list of all surveys available to the user
}

