% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{get_uncoercibles}
\alias{get_uncoercibles}
\title{get_uncoercibles returns values that should be kept as characters.}
\usage{
get_uncoercibles(dataDF)
}
\arguments{
\item{dataDF}{}
}
\description{
get_uncoercibles returns values that should be kept as characters.
}
