% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{longen_psych_wide}
\alias{longen_psych_wide}
\title{Longen a wide psych-scored scale}
\usage{
longen_psych_wide(psychMat, scale_name = "scale", id_colname = "id")
}
\arguments{
\item{psychMat}{a matrix from the \code{$scores} element of a \code{psych} object.}

\item{scale_name}{use to specify the name of the scale.}

\item{id_colname}{use to set the column name of for the ids (taken from the 
rownames of the \code{psychMat}).}
}
\value{
a long data frame with an id column defined by id_colname, 'scale_name', 
'scored_scale', and 'score'
}
\description{
Longen a wide psych-scored scale
}
