% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{make_nice_scale_fname}
\alias{make_nice_scale_fname}
\title{Make nice scale filename}
\usage{
make_nice_scale_fname(scale_name)
}
\arguments{
\item{scale_name}{Name of the scale (a character)}
}
\value{
A character that would be nice to compose into a filename.
}
\description{
Make nice scale filename
}
