% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_psqi.R
\name{psqi_sleep_wake_diff}
\alias{psqi_sleep_wake_diff}
\title{Difference between sleep and wake times}
\usage{
psqi_sleep_wake_diff(timesleep, timewake)
}
\arguments{
\item{timesleep}{Object of class POSIXct on same date as \code{timewake}.}

\item{timewake}{Object of class POSIXct on same date as \code{timesleep}.}
}
\value{
Time elapsed between \code{timesleep} and \code{timewake}.
}
\description{
Difference between sleep and wake times
}
