% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_psqi.R
\name{psqi_to_time_date}
\alias{psqi_to_time_date}
\title{PSQI time to POSIX}
\usage{
psqi_to_time_date(psqi_time, assume = c("bedtime", "waketime"))
}
\arguments{
\item{psqi_time}{character time string formatted as 'H', 'H:M', 'Ip', 'I p', 'I:M p', or 'I:Mp'.}

\item{assume}{For bed times, use 'bedtime'. For wake times, use 'waketime'.}
}
\value{
An object of class POSIXct on an arbitrary date.
}
\description{
PSQI time to POSIX
}
