% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{recode_responses}
\alias{recode_responses}
\title{Recode repsonses}
\usage{
recode_responses(dataDF, recoding_rubric)
}
\arguments{
\item{dataDF}{A long data frame.}

\item{recoding_rubric}{A recoding rubric with columns `column_name`,
`response`, and `score`}
}
\value{
the data frame passed to it, with each value in `value` replaced with
  the recoded value in `score` from the template.
}
\description{
Recodes responses according to a rubric.
}
