% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{reverse_score}
\alias{reverse_score}
\title{Reverse score}
\usage{
reverse_score(
  item_values,
  min = stop("Must specify min and max score."),
  max = stop("Must specify min and max score.")
)
}
\arguments{
\item{item_values}{item_values}

\item{min}{min}

\item{max}{max}
}
\description{
Reverse score
}
