% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{score_items}
\alias{score_items}
\title{Score items}
\usage{
score_items(item_values, scoring_methods, na.rm = F, mean.na.rm = T,
  scale_name = "", scored_scale = "", SID = "")
}
\arguments{
\item{item_values}{item_values}

\item{scoring_methods}{scoring_methods}

\item{na.rm}{na.rm}

\item{mean.na.rm}{mean.na.rm}

\item{scale_name}{scale_name}

\item{scored_scale}{scored_scale}
}
\description{
Score items
}
