% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pds.R
\name{score_pdss}
\alias{score_pdss}
\title{Score pubertal development scale}
\usage{
score_pdss(
  long_survey_data_filtered,
  gender_mix = "mf",
  gendercode = c(m = 1, f = 0)
)
}
\arguments{
\item{long_survey_data_filtered}{data}

\item{gender_mix}{gender_mix}

\item{gendercode}{gendercode}
}
\value{
different version of PDS
}
\description{
Score pubertal development scale
}
