% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_psqi.R
\name{score_psqi}
\alias{score_psqi}
\title{Title}
\usage{
score_psqi(dataDF, pid_col = "ID")
}
\arguments{
\item{dataDF}{A long data frame with columns named: [PID col], item, and value}

\item{pid_col}{The name of the participant ID column.}
}
\value{
A data frame of scored scales with columns scale_name, scored_scale, [PID col], 
score, n_items, n_missing, method.
}
\description{
Title
}
