% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_psqi.R
\name{score_psqi_2}
\alias{score_psqi_2}
\title{Score PSQI Item 2}
\usage{
score_psqi_2(psqi_2)
}
\arguments{
\item{psqi_2}{character that can be coerced to number after text is stripped.}
}
\value{
the corresponding score for this item.
}
\description{
This function assumes any value is in minutes.
}
