% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_psqi.R
\name{score_psqi_4}
\alias{score_psqi_4}
\title{Score PSQI item 4}
\usage{
score_psqi_4(psqi_4, return_value = "score")
}
\arguments{
\item{psqi_4}{duration in hours. May be in the form 'H:M' or as a float like '6.5'.}

\item{return_value}{'score' returns the corresponding item score. 'hours' returns
numeric number of hours.}
}
\value{
score or number of hours.
}
\description{
Score PSQI item 4
}
