% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_psqi.R
\name{score_psqi_component_4}
\alias{score_psqi_component_4}
\title{Score PSQI component 4}
\usage{
score_psqi_component_4(sleepbedratio)
}
\arguments{
\item{sleepbedratio}{A ratio in \code{[0, Inf)}.}
}
\value{
The corresponding PSQI score for component 4.
}
\description{
Score PSQI component 4
}
