% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_psqi.R
\name{score_psqi_component_5}
\alias{score_psqi_component_5}
\title{Score PSQI component 5}
\usage{
score_psqi_component_5(psqi_5_items)
}
\arguments{
\item{psqi_5_items}{A vector of the requisite PSQI sub-items for item 5.}
}
\value{
The corresponding PSQI score for component 5.
}
\description{
Score PSQI component 5
}
