% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{score_questionnaire}
\alias{score_questionnaire}
\title{Score questionnaire}
\usage{
score_questionnaire(dataDF, rubricsDF, psych = FALSE, ...)
}
\arguments{
\item{dataDF}{A data.frame as returned by \code{\link{get_survey_data}}}

\item{rubricsDF}{A data.frame as returned by \code{\link{get_rubrics}}}

\item{psych}{A flag specifying whether \code{\link[psych]{psych-package}} should be used for scoring. Default is \code{FALSE} but you should probably set it to \code{TRUE} unless you're using this for a DSN lab project}

\item{...}{Other arguments passed to one of the scoring engines.}
}
\value{
A long-form data.frame with scale scores.
}
\description{
Score questionnaire
}
