% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{score_questionnaire_dsn}
\alias{score_questionnaire_dsn}
\title{Score questionnaire dsn}
\usage{
score_questionnaire_dsn(dataDF, rubricsDF)
}
\arguments{
\item{dataDF}{dataDF}

\item{rubricsDF}{rubricsDF}
}
\description{
Score questionnaire dsn
}
