% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{score_questionnaire_psych}
\alias{score_questionnaire_psych}
\title{score questionnaire psych}
\usage{
score_questionnaire_psych(dataDF, rubricsDF, scale_name = NULL,
  return_with_data = FALSE)
}
\arguments{
\item{dataDF}{dataDF}

\item{rubricsDF}{rubricsDF}

\item{scale_name}{scale_name}

\item{return_with_data}{return_with_data}
}
\description{
score questionnaire psych
}
