% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_psqi.R
\name{score_single_psqi_row}
\alias{score_single_psqi_row}
\title{Score a single PSQI questionnaire}
\usage{
score_single_psqi_row(psqi_data)
}
\arguments{
\item{psqi_data}{A row of PSQI data (see details)}
}
\value{
A data frame with columns named scored_scale, 
score, n_items, n_missing, method.
}
\description{
Score a single PSQI questionnaire
}
\details{
The vector of PSQI items must have the following names: 
PSQI_1_1_TEXT, PSQI_1_2_TEXT, PSQI_2_1_TEXT, PSQI_2_2_TEXT,
PSQI_3_1_TEXT, PSQI_3_2_TEXT, PSQI_4_1_TEXT, PSQI_4_2_TEXT,
PSQI_5_a, PSQI_5_b, PSQI_5_c, PSQI_5_d, PSQI_5_e, PSQI_5_f,
PSQI_5_g, PSQI_5_h, PSQI_5_i, PSQI_6, PSQI_7, PSQI_8, PSQI_9
}
