% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{score_step_one_and_two}
\alias{score_step_one_and_two}
\title{Score step one and two}
\usage{
score_step_one_and_two(dataDF, rubricsDF)
}
\arguments{
\item{dataDF}{A data.frame as returned by \code{\link{get_survey_data}}}

\item{rubricsDF}{A data.frame as returned by \code{\link{get_rubrics}}}
}
\value{
A long-form data.frame with scale scores.
}
\description{
Score step one and two
}
