% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{transform_scores}
\alias{transform_scores}
\title{transform scores}
\usage{
transform_scores(item_values, transformation, min = NA, max = NA)
}
\arguments{
\item{item_values}{item_values}

\item{transformation}{transformation}

\item{min}{min}

\item{max}{max}
}
\description{
transform scores
}
