% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{widen_qualtrics_long}
\alias{widen_qualtrics_long}
\title{Widen Qualtrics Long}
\usage{
widen_qualtrics_long(dataDF, scale_names)
}
\arguments{
\item{dataDF}{a scored data frame}

\item{scale_names}{a vector of scale names to be matched exactly}
}
\value{
a list with \code{scores}, a wide format data frame with scale scores,
and \code{data_quality}, a wide format data frame with number if missing and used items
for each scale.
}
\description{
Widen Qualtrics Long
}
