% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{write_widened_scored_scale}
\alias{write_widened_scored_scale}
\title{Write widened scored scale}
\usage{
write_widened_scored_scale(dataDF, scale_names = NULL, dir_name = NULL,
  file_name = NULL, metadata = NULL)
}
\arguments{
\item{dataDF}{A long data frame.}

\item{scale_names}{A character vector of scale names to widen and write.}

\item{dir_name}{The output directory to save csv files.}

\item{file_name}{A custom file name.}

\item{metadata}{List of key-value pairs that will be added as
column name - column values to the returned data frames.}
}
\description{
Write widened scored scale
}
