% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{creds_from_file}
\alias{creds_from_file}
\title{Get authentication}
\usage{
creds_from_file(creds_yaml = "credentials.yaml")
}
\arguments{
\item{creds_yaml}{The .yaml file that contains the user's credentials}
}
\description{
Instructions here: https://survey.qualtrics.com/WRAPI/ControlPanel/docs.php#authentication_2.3
}
