% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{get_survey_data}
\alias{get_survey_data}
\title{Get survey data}
\usage{
get_survey_data(surveysDF, pid_cols = "ID", ...)
}
\arguments{
\item{surveysDF}{A data.frame as returned by the \code{\link{get_surveys}}.}

\item{...}{Not used.}

\item{pid_col}{A character vector with the column names specifying user ID
and other column names to use as ID variables in data-lengthening.}
}
\value{
Returns a long format data.frame of survey data with names ["SID
  variable names", ...], "item", "value", "survey_name"
}
\description{
Get survey data
}
