% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_requests.R
\name{get_survey_responses}
\alias{get_survey_responses}
\title{Return a data.frame of the survey responses for a given survey}
\usage{
get_survey_responses(creds, user = creds$user, token = creds$token,
  surveyid, format = "JSON", version = "2.4", otherArgs = NULL)
}
\arguments{
\item{creds}{Qualtrics credentials. A list with values for user and token.}

\item{user}{Qualtrics username.}

\item{token}{Qualtrics API token.}

\item{surveyid}{Qualtrics survey ID.}

\item{format}{The format of the requested data. Default is JSON.}

\item{version}{The API version to use. Defaults to the latest version, 2.4.}

\item{otherArgs}{A `list` of other arguments to provide the API.}
}
\description{
Return a data.frame of the survey responses for a given survey
}
